Update ffrkMasteryScraper.py
In this version I correctly edited all the list of threadIds and commentIds to be lists of lists, so that parseMasterySubmission() works correctly when it's called in each chunk of the main code block.

I also updated the main code block section that runs over wodinThreadIds so it will support multiple threadIds when there are eventually new surveys added. (Did NOT update the main code block section that runs over wodinCommentIds, since it's unclear how new surveys will eventually be added for those WOdin forms - as comments on a single thread, or as individual threads for each one? we'll have to wait and see)

This version still includes updates to allow for LBGs, ADSBs, and a couple more catches for GSBs.

And I checked BOTH of the PostText.txt outputs this time! :D

Let me know what you think!